WIP - make cmake independent from pip
With the changes introduces in PEP 668, pip discourages users to do a system install of python dependencies. In this context, it is imperative that we remove dependency on pip for installing pybind11 and other modules.

Because of PEP 668, the new thumb rule is to keep cmake independent of pip-based installation of C++ libraries, including pybind11. Instead, `FetchContent_Declare` or `ExternalProject_Add` cmake commands should be used to download and install from source/ unpack compressed binaries.

We would continue to maintain a separate installation of all dependencies, including boost, pybind11, and catch2. We think that this step is important to maintain version compatibility while not impacting(breaking) any system packages. In a sense, our decision is driven by the same factors that make docker images useful.

Following action items arise:
1. Remove pip-based installations from cmake scripts, replacing them by FetchContent_Declare or ExternalProject_Add
2. Create requirement files based on the LAPKT component being built. e.g. C++-Lib-Only, With-Python-Package, LAPKT-documentation
3. Document readmes instructing the developers on how to build C++-Lib-Only, With-Python-Package, and LAPKT-documentation